Failed to get list of releases of private repository
Getting releases from a private repository using the GITHUB_TOKEN in the action session running in repo deployments does nit give access. Created a PAT for atvgitops in it;s account and using that value as a secret in the atvanzt organization. This should give access to the releases, I hope!!